feat: split public workspace crates
