+++ Cargo.toml
@@ -25,14 +25,24 @@ exclude = [
     "books/**",       # mdBook manuscript + 2.5MB local Mermaid asset; repo/docs only, not runtime
 ]
 
+[workspace]
+members = [
+    ".",
+    "crates/mempal-agent-memory",
+    "crates/mempal-embed",
+    "crates/mempal-mcp-protocol",
+    "crates/mempal-search-core",
+]
+resolver = "2"
+
 [[bin]]
 name = "mempal"
 path = "src/main.rs"
 
 [features]
 default = ["model2vec"]
-model2vec = ["dep:model2vec-rs"]
-onnx = ["dep:ort", "dep:tokenizers"]
+model2vec = ["mempal-embed/model2vec"]
+onnx = ["mempal-embed/onnx"]
 rest = ["dep:axum", "dep:tower", "dep:tower-http", "model2vec"]
 
 [dependencies]
@@ -43,8 +53,10 @@ bimap = "0.6"
 clap = { version = "4", features = ["derive"] }
 ignore = "0.4"
 jieba-rs = "0.8"
-model2vec-rs = { version = "0.1", optional = true }
-ort = { version = "2.0.0-rc.9", optional = true }
+mempal-agent-memory = { version = "0.8.0", path = "crates/mempal-agent-memory" }
+mempal-embed = { version = "0.8.0", path = "crates/mempal-embed", default-features = false }
+mempal-mcp-protocol = { version = "0.8.0", path = "crates/mempal-mcp-protocol" }
+mempal-search-core = { version = "0.8.0", path = "crates/mempal-search-core" }
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 rmcp = { version = "1.3.0", features = ["server", "macros", "schemars", "transport-async-rw", "transport-io"] }
 rusqlite = { version = "0.37", features = ["bundled"] }
@@ -54,7 +66,6 @@ sha2 = "0.10"
 sqlite-vec = "0.1.9"
 tempfile = "3"
 thiserror = "2"
-tokenizers = { version = "0.22", optional = true }
 tokio = { version = "1", features = ["full"] }
 toml = "0.8"
 tower = { version = "0.5", optional = true }
