+++ crates/mempal-agent-memory/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "mempal-agent-memory"
+version = "0.8.0"
+edition = "2024"
+license = "MIT"
+description = "Reusable agent-memory domain types and anchor helpers for mempal."
+repository = "https://github.com/ZhangHanDong/mempal"
+homepage = "https://github.com/ZhangHanDong/mempal"
+documentation = "https://docs.rs/mempal-agent-memory"
+rust-version = "1.85"
+readme = "../../README.md"
+keywords = ["memory", "agents", "knowledge", "mempal"]
+categories = ["development-tools"]
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+thiserror = "2"
