+++ crates/mempal-agent-memory/src/anchor.rs
