+++ crates/mempal-agent-memory/src/lib.rs
@@ -1,0 +1,4 @@
+#![warn(clippy::all)]
+
+pub mod anchor;
+pub mod types;
