+++ crates/mempal-agent-memory/src/types.rs
