+++ crates/mempal-embed/src/api.rs
