+++ crates/mempal-embed/src/model2vec.rs
