+++ crates/mempal-embed/src/onnx.rs
