+++ src/core/mod.rs
@@ -1,9 +1,9 @@
 #![warn(clippy::all)]
 
-pub mod anchor;
 pub mod config;
 pub mod db;
 pub mod phase3;
 pub mod protocol;
-pub mod types;
 pub mod utils;
+
+pub use mempal_agent_memory::{anchor, types};
