+++ src/embed/mod.rs
@@ -1,108 +1,62 @@
 #![warn(clippy::all)]
 
-use std::path::PathBuf;
+use async_trait::async_trait;
 
-use thiserror::Error;
+use crate::core::config::Config;
 
-pub mod api;
-pub mod factory;
 #[cfg(feature = "model2vec")]
-pub mod model2vec;
+pub use mempal_embed::model2vec;
 #[cfg(feature = "onnx")]
-pub mod onnx;
+pub use mempal_embed::onnx;
+pub use mempal_embed::{EmbedError, Embedder, EmbedderFactory, Result, api};
 
-pub use factory::{ConfiguredEmbedderFactory, EmbedderFactory};
+/// Default embedding dimensions used by the API backend when config omits one.
+pub const DEFAULT_API_DIMENSIONS: usize = 384;
 
-pub type Result<T> = std::result::Result<T, EmbedError>;
-
-#[derive(Debug, Error)]
-pub enum EmbedError {
-    #[error("failed to create model directory {path}")]
-    CreateModelDir {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-    #[error("failed to check whether {path} exists")]
-    CheckPathExists {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-    #[error("failed to download {url}")]
-    Download {
-        url: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("download returned error status for {url}")]
-    DownloadStatus {
-        url: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("failed to read download body from {url}")]
-    ReadDownloadBody {
-        url: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("failed to write {path}")]
-    WriteFile {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-    #[error("failed to rename {from} to {to}")]
-    RenameFile {
-        from: PathBuf,
-        to: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-    #[error("failed to initialize ONNX session builder: {0}")]
-    SessionBuilder(String),
-    #[error("failed to load ONNX model from {path}: {message}")]
-    LoadModel { path: PathBuf, message: String },
-    #[error("tokenizer error: {0}")]
-    Tokenizer(String),
-    #[error("embedding runtime error: {0}")]
-    Runtime(String),
-    #[error("embedding worker panicked")]
-    WorkerPanic(#[source] tokio::task::JoinError),
-    #[error("failed to call embedding endpoint {endpoint}")]
-    HttpRequest {
-        endpoint: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("embedding endpoint returned error status {endpoint}")]
-    HttpStatus {
-        endpoint: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("failed to decode embedding response from {endpoint}")]
-    DecodeResponse {
-        endpoint: String,
-        #[source]
-        source: reqwest::Error,
-    },
-    #[error("invalid embedding response: {0}")]
-    InvalidResponse(String),
-    #[error("embedding endpoint returned no vectors")]
-    EmptyVectors,
-    #[error(
-        "embedding endpoint returned vectors with unexpected dimensions; expected {expected}, got {actual}"
-    )]
-    InvalidDimensions { expected: usize, actual: usize },
-    #[error("unsupported embed backend: {0}")]
-    UnsupportedBackend(String),
+#[derive(Clone)]
+pub struct ConfiguredEmbedderFactory {
+    config: Config,
 }
 
-#[async_trait::async_trait]
-pub trait Embedder: Send + Sync {
-    async fn embed(&self, texts: &[&str]) -> Result<Vec<Vec<f32>>>;
-    fn dimensions(&self) -> usize;
-    fn name(&self) -> &str;
+impl ConfiguredEmbedderFactory {
+    pub fn new(config: Config) -> Self {
+        Self { config }
+    }
+}
+
+#[async_trait]
+impl EmbedderFactory for ConfiguredEmbedderFactory {
+    async fn build(&self) -> Result<Box<dyn Embedder>> {
+        match self.config.embed.backend.as_str() {
+            #[cfg(feature = "model2vec")]
+            "model2vec" => {
+                let model_id = self
+                    .config
+                    .embed
+                    .model
+                    .as_deref()
+                    .unwrap_or("minishlab/potion-multilingual-128M");
+                Ok(Box::new(model2vec::Model2VecEmbedder::new(model_id)?))
+            }
+            #[cfg(not(feature = "model2vec"))]
+            "model2vec" => Err(EmbedError::UnsupportedBackend("model2vec".to_string())),
+            #[cfg(feature = "onnx")]
+            "onnx" => Ok(Box::new(onnx::OnnxEmbedder::new_or_download().await?)),
+            #[cfg(not(feature = "onnx"))]
+            "onnx" => Err(EmbedError::UnsupportedBackend("onnx".to_string())),
+            "api" => Ok(Box::new(api::ApiEmbedder::new(
+                self.config
+                    .embed
+                    .api_endpoint
+                    .clone()
+                    .unwrap_or_else(|| "http://localhost:11434/api/embeddings".to_string()),
+                self.config.embed.api_model.clone(),
+                self.config
+                    .embed
+                    .dimensions
+                    .unwrap_or(DEFAULT_API_DIMENSIONS),
+            ))),
+            other => Err(EmbedError::UnsupportedBackend(other.to_string())),
+        }
+    }
 }
